replace DBContext and Migrations to separate project + added connection to Azure database
