+++ LiveChat/LiveChat.csproj
@@ -176,11 +176,6 @@
     <Compile Include="Global.asax.cs">
       <DependentUpon>Global.asax</DependentUpon>
     </Compile>
-    <Compile Include="Migrations\201803102027538_Initial.cs" />
-    <Compile Include="Migrations\201803102027538_Initial.Designer.cs">
-      <DependentUpon>201803102027538_Initial.cs</DependentUpon>
-    </Compile>
-    <Compile Include="Migrations\Configuration.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Startup.cs" />
   </ItemGroup>
@@ -254,11 +249,10 @@
     <Content Include="Scripts\jquery-1.10.2.min.map" />
   </ItemGroup>
   <ItemGroup>
-    <EmbeddedResource Include="Migrations\201803102027538_Initial.resx">
-      <DependentUpon>201803102027538_Initial.cs</DependentUpon>
-    </EmbeddedResource>
-  </ItemGroup>
-  <ItemGroup>
+    <ProjectReference Include="..\DB\DB.csproj">
+      <Project>{3bd6dce8-0fe9-4cce-aacf-c45580fc3608}</Project>
+      <Name>DB</Name>
+    </ProjectReference>
     <ProjectReference Include="..\Models\Models.csproj">
       <Project>{63b40d81-f270-4911-bd45-493405490b97}</Project>
       <Name>Models</Name>
