+++ LiveChat/Web.config
@@ -11,7 +11,10 @@
   <connectionStrings>
     <!--add name="DefaultConnection" connectionString="Data Source=(LocalDb)\MSSQLLocalDB;AttachDbFilename=|DataDirectory|\aspnet-LiveChat-20180306114259.mdf;Initial Catalog=aspnet-LiveChat-20180306114259;Integrated Security=True"
       providerName="System.Data.SqlClient" /-->
-    <add name="DefaultConnection" connectionString="Data Source=WIN-AFU633JOFGJ\SQLEXPRESS;          Initial Catalog=LiveChat;          Integrated Security=True" providerName="System.Data.SqlClient" />
+    <!--add name="DefaultConnection" connectionString="Data Source=WIN-AFU633JOFGJ\SQLEXPRESS;          Initial Catalog=LiveChat;          Integrated Security=True" providerName="System.Data.SqlClient" /-->
+    <!--add name="DefaultConnection" connectionString="Data Source=win-afu633jofgj.database.windows.net;Initial Catalog=LiveChat;Integrated Security=False;User ID=woodpecker;Password=Aa123456;Connect Timeout=60;Encrypt=False;TrustServerCertificate=True;ApplicationIntent=ReadWrite;MultiSubnetFailover=False" /-->
+    <!--add name="DefaultConnection" connectionString="Server=tcp:win-afu633jofgj.database.windows.net,1433;Initial Catalog=LiveChat;Persist Security Info=False;User ID={your_username};Password={your_password};MultipleActiveResultSets=False;Encrypt=True;TrustServerCertificate=False;Connection Timeout=30;" providerName="System.Data.EntityClient"/-->
+    <add name="DefaultConnection" connectionString="Server=tcp:win-afu633jofgj.database.windows.net,1433;Database=LiveChat;User ID=woodpecker;Password=Aa123456;Trusted_Connection=False;Encrypt=True;Connection Timeout=30;" providerName="System.Data.SqlClient"/>
     <!--add name="DefaultConnection" 
          connectionString=
          "Data Source=WIN-AFU633JOFGJ\SQLEXPRESS;
